[modify] 処理を調整
